refactor(nbd-cow): split device pool actor and state
